add new components

+++ src/layout/Navbar.vue
@@ -13,22 +13,36 @@
       <div class="item item-lock">
         <i class="iconfont icon-lock"></i>
       </div>
-      <div class="item item-set">
-        <i class="iconfont icon-setting"></i>
-      </div>
+      <el-popover v-model="sysVisible" placement="bottom" trigger="hover" popper-class="user-popover">
+        <div class="user-pop-box">
+          <div class="user-item el-select-dropdown__item" @mouseenter="showLang = true" @mouseleave="showLang = false">
+            <i class="iconfont el-icon-collection-tag"></i> <span>{{ $t('system.langSlt') }}</span>
+            <div v-show="showLang" class="lang-pop">
+              <div class="el-select-dropdown__item" @click="changeLang('zh')">简体中文</div>
+              <div class="el-select-dropdown__item" @click="changeLang('en')">English</div>
+            </div>
+          </div>
+          <div class="user-item el-select-dropdown__item">
+            <i class="iconfont el-icon-connection"></i> <span>{{ $t('system.themeSlt') }}</span>
+          </div>
+        </div>
+        <div class="item item-set" slot="reference">
+          <i class="iconfont icon-setting"></i>
+        </div>
+      </el-popover>
       <el-popover v-model="userVisible" placement="bottom" trigger="hover" popper-class="user-popover">
         <div class="user-pop-box">
           <div class="user-item el-select-dropdown__item">
-            <i class="iconfont icon-edit"></i> <span>资料设置</span>
+            <i class="iconfont icon-edit"></i> <span>{{ $t('system.profileSet') }}</span>
           </div>
           <div class="user-item el-select-dropdown__item">
-            <i class="iconfont icon-yzm"></i> <span>修改密码</span>
+            <i class="iconfont icon-yzm"></i> <span>{{ $t('system.editPassword') }}</span>
           </div>
           <div class="user-item el-select-dropdown__item">
-            <i class="iconfont icon-audit"></i> <span>登录记录</span>
+            <i class="iconfont icon-audit"></i> <span>{{ $t('system.loginRecord') }}</span>
           </div>
           <div class="user-item el-select-dropdown__item" @click="loginOut">
-            <i class="iconfont icon-exit"></i> <span>退出系统</span>
+            <i class="iconfont icon-exit"></i> <span>{{ $t('system.exitSys') }}</span>
           </div>
         </div>
         <div class="item item-user" slot="reference">
@@ -40,22 +54,38 @@
 </template>
 
 <script>
+import CommonUtil from "@/utils/commonUtil";
+
 export default {
   name: 'navBar',
   data() {
     return {
+      language: 'zh',
+      showLang: false,
       showPop: false,
       searchMenu: '',
       userVisible: false,
+      sysVisible: false,
       loading: false
     }
   },
   methods: {
     loginOut() {
-      this.userVisible = true
-      window.localStorage.clear()
-      this.$router.push('/login')
-      // this.$message.success('用户注销')
+      this.$confirm('此操作将退出系统, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        window.localStorage.clear()
+        this.$router.push('/login')
+      })
+    },
+    changeLang(type) {
+      const lang = CommonUtil.getStorageItem('language')
+      if (lang !== type) {
+        CommonUtil.setStorageItem('language', type)
+        window.location.reload()
+      }
     }
   }
 }
@@ -91,6 +121,7 @@ export default {
     display: flex;
     justify-content: center;
     align-items: center;
+    position: relative;
     .item{
       width: 40px;
       height: 50px;
@@ -127,6 +158,13 @@ export default {
         top: 5px;
         right: -6px;
       }
+    }
+  }
+  .i18ns{
+    width: 130px;
+    /deep/ .el-input__inner{
+      height: 28px;
+      line-height: 28px;
     }
   }
 }
